redefine types,  rearrange methods and add docstrings to describe why the methods are needed
